+++ Microsoft.Azure.Cosmos/src/CosmosClientOptionsFeatures.cs
@@ -1,0 +1,15 @@
+﻿// ------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.  All rights reserved.
+// ------------------------------------------------------------
+
+namespace Microsoft.Azure.Cosmos
+{
+    using System;
+
+    [Flags]
+    internal enum CosmosClientOptionsFeatures
+    {
+        NoFeatures = 0,
+        AllowBulkExecution = 1
+    }
+}