+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/Json/JsonWriterTests.cs
@@ -1,4 +1,4 @@
-﻿namespace Microsoft.Azure.Cosmos.NetFramework.Tests.Json
+﻿namespace Microsoft.Azure.Cosmos.Tests.Json
 {
     using System;
     using System.Collections.Generic;
@@ -412,8 +412,7 @@
                 // remove formatting on the json and also replace "/" with "\/" since newtonsoft is dumb.
                 expectedString = Newtonsoft.Json.Linq.JToken
                     .Parse(expectedString)
-                    .ToString(Newtonsoft.Json.Formatting.None)
-                    .Replace("/", @"\/");
+                    .ToString(Newtonsoft.Json.Formatting.None);
 
                 byte[] binaryOutput =
                 {
@@ -879,7 +878,6 @@
                 new Tuple<string, string>(@"\t", "\t"),
                 new Tuple<string, string>(@"\""", "\""),
                 new Tuple<string, string>(@"\\", @"\"),
-                new Tuple<string, string>(@"\/", "/"),
             };
 
             foreach (Tuple<string, string> escapeCharacter in escapeCharacters)
@@ -987,7 +985,7 @@
                 // Whitespace characters have special escaping
                 if (!escapeCharacters.Contains((char)controlCharacter))
                 {
-                    string expectedString = "\"" + "\\u" + "00" + controlCharacter.ToString("x2") + "\"";
+                    string expectedString = "\"" + "\\u" + "00" + controlCharacter.ToString("X2") + "\"";
 
                     JsonToken[] tokensToWrite =
                     {
@@ -1056,7 +1054,7 @@
                 binaryOutputBuilder.Add(new byte[] { BinaryFormat, JsonBinaryEncoding.TypeMarker.Object1ByteLength });
 
                 List<byte[]> elements = new List<byte[]>();
-                elements.Add(new byte[] { (byte)(JsonBinaryEncoding.TypeMarker.UserString1ByteLengthMin) });
+                elements.Add(new byte[] { (byte)JsonBinaryEncoding.TypeMarker.UserString1ByteLengthMin });
                 elements.Add(new byte[] { JsonBinaryEncoding.TypeMarker.LiteralIntMin + 10 });
                 elements.Add(new byte[] { (byte)(JsonBinaryEncoding.TypeMarker.UserString1ByteLengthMin + 1) });
                 elements.Add(new byte[] { (byte)(JsonBinaryEncoding.TypeMarker.EncodedStringLengthMin + "example glossary".Length), 101, 120, 97, 109, 112, 108, 101, 32, 103, 108, 111, 115, 115, 97, 114, 121 });
@@ -1175,7 +1173,7 @@
                 elements.Add(new byte[] { (byte)(JsonBinaryEncoding.TypeMarker.SystemString1ByteLengthMin + 12) });
                 elements.Add(new byte[] { (byte)(JsonBinaryEncoding.TypeMarker.EncodedStringLengthMin + "7029d079-4016-4436-b7da-36c0bae54ff6".Length), 55, 48, 50, 57, 100, 48, 55, 57, 45, 52, 48, 49, 54, 45, 52, 52, 51, 54, 45, 98, 55, 100, 97, 45, 51, 54, 99, 48, 98, 97, 101, 53, 52, 102, 102, 54 });
 
-                elements.Add(new byte[] { (byte)(JsonBinaryEncoding.TypeMarker.UserString1ByteLengthMin) });
+                elements.Add(new byte[] { (byte)JsonBinaryEncoding.TypeMarker.UserString1ByteLengthMin });
                 elements.Add(new byte[] { JsonBinaryEncoding.TypeMarker.NumberDouble, 0x98, 0x8B, 0x30, 0xE3, 0xCB, 0x45, 0xC8, 0x3F });
 
                 elements.Add(new byte[] { (byte)(JsonBinaryEncoding.TypeMarker.UserString1ByteLengthMin + 1) });
@@ -1598,7 +1596,7 @@
 
                 JsonToken[] tokensToWrite =
                 {
-                    JsonToken.Binary(new List<byte>())
+                    JsonToken.Binary(new byte[]{ })
                 };
 
                 this.VerifyWriter(tokensToWrite, expectedStringOutput);
@@ -1607,7 +1605,7 @@
 
             {
                 // Binary 1 Byte Length
-                IReadOnlyList<byte> binary = Enumerable.Range(0, 25).Select(x => (byte)x).ToList();
+                byte[] binary = Enumerable.Range(0, 25).Select(x => (byte)x).ToArray();
                 string expectedStringOutput = $"B{Convert.ToBase64String(binary.ToArray())}";
                 byte[] expectedBinaryOutput;
                 unchecked
@@ -1658,37 +1656,6 @@
                 {
                     System.Threading.Thread.CurrentThread.CurrentCulture = cultureInfo;
 
-                    Encoding[] encodings =
-                    {
-                        Encoding.UTF8,
-                        Encoding.Unicode,
-                        Encoding.UTF32,
-                    };
-
-                    foreach (Encoding encoding in encodings)
-                    {
-                        // Create through encoding API
-                        IJsonWriter jsonWriterWithEncoding = JsonWriter.Create(encoding);
-                        if (expectedString != null)
-                        {
-                            this.VerifyWriter(
-                                jsonWriterWithEncoding, 
-                                tokensToWrite, 
-                                encoding.GetBytes(expectedString), 
-                                JsonSerializationFormat.Text, 
-                                expectedException);
-                        }
-                        else
-                        {
-                            this.VerifyWriter(
-                                jsonWriterWithEncoding, 
-                                tokensToWrite, 
-                                null, 
-                                JsonSerializationFormat.Text, 
-                                expectedException);
-                        }
-                    }
-
                     // Create through serializtion api
                     IJsonWriter jsonWriter = JsonWriter.Create(JsonSerializationFormat.Text);
                     byte[] expectedOutput;
@@ -1702,10 +1669,10 @@
                     }
 
                     this.VerifyWriter(
-                        jsonWriter, 
-                        tokensToWrite, 
-                        expectedOutput, 
-                        JsonSerializationFormat.Text, 
+                        jsonWriter,
+                        tokensToWrite,
+                        expectedOutput,
+                        JsonSerializationFormat.Text,
                         expectedException);
                 }
             }
@@ -1822,8 +1789,8 @@
                             break;
 
                         case JsonTokenType.Binary:
-                            IReadOnlyList<byte> binaryValue = (token as JsonBinaryToken).Value;
-                            jsonWriter.WriteBinaryValue(binaryValue);
+                            ReadOnlyMemory<byte> binaryValue = (token as JsonBinaryToken).Value;
+                            jsonWriter.WriteBinaryValue(binaryValue.Span);
                             break;
 
                         case JsonTokenType.NotStarted:
@@ -1841,11 +1808,18 @@
 
             if (expectedException == null)
             {
-                byte[] result = jsonWriter.GetResult();
-                Assert.IsTrue(expectedOutput.SequenceEqual(result),
-                    string.Format("Expected : {0}, Actual :{1}",
-                    string.Join(", ", expectedOutput),
-                    string.Join(", ", result)));
+                byte[] result = jsonWriter.GetResult().ToArray();
+                if (jsonSerializationFormat == JsonSerializationFormat.Text)
+                {
+                    Assert.AreEqual(Encoding.UTF8.GetString(expectedOutput), Encoding.UTF8.GetString(result));
+                }
+                else
+                {
+                    Assert.IsTrue(expectedOutput.SequenceEqual(result),
+                        string.Format("Expected : {0}, Actual :{1}",
+                        string.Join(", ", expectedOutput),
+                        string.Join(", ", result)));
+                }
             }
         }
     }